use long to avoid overflow exceptions for tours with many trips

+++ src/main/java/ch/ethz/matsim/discrete_mode_choice/model/mode_chain/ModeChainGenerator.java
@@ -11,5 +11,5 @@ import java.util.List;
  * @author sebhoerl
  */
 public interface ModeChainGenerator extends Iterator<List<String>> {
-	int getNumberOfAlternatives();
+	long getNumberOfAlternatives();
 }
